Add support for AWS STS
WAL-E now accepts AWS_SESSION_TOKEN in order to use the temporary
credentials vended by AWS STS.

Also included is a sensible policy one might use with STS.  A slight
adjustment to backup-list has been made so that it can operate while
in use with this policy.

Use of STS requires a newer version of boto.  The first version
thought to accept a security_token parameter is 2.6.